levels/world2: Add "This Forest is Haunted"
"This Forest is Haunted" is an additional new Ghost Forest level. It
features a second exit hidden in a secret area.
It can be used as a bonus level or in a separate Ghost Forest path on
the worldmap.

Ref: SuperTux/data#69 ("levels/world2: Add "Bye Bye Forest" + "This
Forest Is Haunted"")
[max@teufelsnetz.com: taken from 31a2a457b7924f66af3794f71a7af66d175cbba5,
split off "This Forest is Haunted", and extended commit message]
Reviewed-by: Max Teufel <max@teufelsnetz.com>
Tested-by: Max Teufel <max@teufelsnetz.com>
Cc: alex jockers <alex.jockers@gmx.de>
Cc: Max Teufel <max@teufelsnetz.com>
Cc: Carsten Wirtz <carstenwirtz@live.de>